Block Scrapper BSC v02
Improved the code, now it's assured that Block Scrapper BSC won't return a ValueError if a random non-BNB value (such as 1 wei) appears in one cell of the rows when reading the column 'BNB Value'. Fortunately, we found this error while testing our previous version of Block Scrapper BSC in the block 9875880, which a Block No. pretty close to our very first tests!