Update Block Scrapper BSC.py
Improved the logic of code, now the user only has to assign manually the Block No. at line 28 to make the program run for the next 100 blocks (including the initial one), and finally export the final dataframe with a dynamic filename that depends on the values of "initial_block" & "final_block" variables.