THIS PROGRAM CAN'T BE STOPPED NOW! Hell yeah
After testing several times the previous version of this program, it was found that VERY OCASSIONALLY the error "TimeoutException" appeared when trying to get the contract_age element from the explorer.bitquery.io, which made the program stop and crash everything until then. This situation rustled my jimmies so hard, that I decided to patch it with a simple combination of a while loop and try and exception blocks, making sure that from this time this "TimeoutException" prank is not going to happen again. It was also added a few new contracts to the lista_negra dictionary