¡Habemus Lotería!
After testing once again the previous version of this program, it was found that a MaxRetryError was raising a lot of times, and it was very weird, initially it was thought that BscScan was making some sort of deep-ass analysis in order to block my IP , fortunately, that was not the case, it was just that the driver was not reinitialized after executing the driver.quit() function. It was also modified the source of information where the data for the initial_block variable are called, just in case BscScan wants to fuck things up. It was also added a pseudo wait for loop at the end of the cycle just to make the process more "human". Anyway, now this program seems to be capable of running itself every day non-stop 100% guaranteed, but that's something we hopefully will find out soon!