Too soon, but not too late ;)
It was realized that BscScan doesn't give a fek about web scrapping, it was tested the previous code all over the BscScan website only, throwing 0 errors, it was also reduced the pseudo wait at the end of the code because it's actually unnecesary lol, the lista_negra was improved with more undesired contract addresses, and also it was added an if statement at 400 line which will export the dataframe_definitiva only if it's not empty. This program seems to be capable of running in automatic mode still, so that's good too.