Minor updates, checked
A few more contracts were added to lista_negra and also its content was reorganized to make it occupy less lines of code. It was also observed that there were very few situations in which a Transaction Id contained an large list of Tokens Transferred (i.e. Dividend Trackers and crap like that, usually more than 6 elements), making this program loop forever in the while z loop at line 224, it was decided to limit the amount of Tokens Transferred by setting it to less or equal to 5 at line 243, hopefully we won't see dividend shitcoins anymore