IMPORTANT IMPROVEMENTS! #MustRead
Firstly, lista_negra got some new contracts to avoid, as always, not so important but you know, still enough to make it the first announcement. Secondly, driver.refresh() method was added to the loops in which is intended to get the information of the following variables: status, Hodlers, contract_age. Believe it or not, these muthafuckas didn't want to get their ass scraped due to some weird-ass problem with the queries when loading the page in which they exist, besides that the loops in which they are called were only ordering the bot to read the information of the variable if it was available, and in case of not, keep waiting, but there were some situations in which these pages were not fully loaded, so it was necessary to F5'd the shit out of them in order to make these loops great again. time.sleep() delays were also added together with these driver.refresh() methods, because we wanted to make sure this time this patch becomes "proof-of-fools". And Thirdly, we increased our initial conditions parameters(Minimum amount of: MORE Hodlers, MORE Transactions based on MORE Hodlers) because we think it's better to get info of projects which are part of the hype train and have a good Hodler base, and a very low slippage (the slowest if possible).

That's pretty it, thanks for reading.